+++ packages/zowe-explorer/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 
 ### Bug fixes
 
+- Fixed issue where the "Disable Validation for Profile" context menu option did not update to "Enable Validation for Profile" after use. [#1897](https://github.com/zowe/vscode-extension-for-zowe/issues/1897)
+
 ## `2.8.0`
 
 ### New features and enhancements
@@ -33,6 +35,7 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 - Fixed issue where a job search query would not expand the session node after it has been filtered.
 - Fixed issue where refreshing views collapsed the respective trees. [#2215](https://github.com/zowe/vscode-extension-for-zowe/issues/2215)
 - Fixed error message when no data sets found that match pattern.
+- Fixed secure credential storage not possible to enable in Theia.
 
 ## `2.7.0`